+++ superwall/src/main/java/com/superwall/sdk/network/API.kt
@@ -6,7 +6,8 @@ import com.superwall.sdk.config.options.SuperwallOptions
 data class Api(
     val hostDomain: String,
     val base: Base,
-    val collector: Collector
+    val collector: Collector,
+    val geo: Geo
 ) {
     companion object {
         const val version1 = "/api/v1/"
@@ -18,7 +19,8 @@ data class Api(
     constructor(networkEnvironment: SuperwallOptions.NetworkEnvironment) : this(
         hostDomain = networkEnvironment.hostDomain,
         base = Base(networkEnvironment),
-        collector = Collector(networkEnvironment)
+        collector = Collector(networkEnvironment),
+        geo = Geo(networkEnvironment)
     )
 
     data class Base(private val networkEnvironment: SuperwallOptions.NetworkEnvironment) {
@@ -31,5 +33,11 @@ data class Api(
         val host: String
             //            get() = "10.0.2.2:9909"
             get() = networkEnvironment.collectorHost
+    }
+
+    data class Geo(private val networkEnvironment: SuperwallOptions.NetworkEnvironment) {
+        val host: String
+            //            get() = "10.0.2.2:9909"
+            get() = "geo-api.superwall.com"
     }
 }